Added kalman prediction separately
It is still not properly added.
We are not using it to actually predict and the values given as
"control" are wrong, we give it corr_vec after it has been adjusted to
step size, when it should be pixel size. infact I'll just fix that right
now.

Moreover I havn't actually tried it with a real setup, only old movies.
Still pretty promising results! We'll see how things go :)